% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{.marginSums}
\alias{.marginSums}
\title{Backport from R 4}
\usage{
.marginSums(x, margin = NULL)
}
\arguments{
\item{x}{object to summarise}

\item{margin}{margin over which to marginalize}
}
\description{
Backport from R 4
}
